fix 000

+++ 12_Special_member_functions/5_Free_store/1_ok/test000.cpp
@@ -27,6 +27,6 @@ void* B::operator new(size_t sz, Arena* p)
 {
   printf("`B::operator new' called\n");
   printf("sz = %d\n", (int)sz);
-  printf("p = %x\n", p);
+  printf("p = %x\n", (int)p);
   return p;
 }
